+++ database/migrations/2023_06_12_174305_create_purchase_details_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->unsignedBigInteger('bill_id');
             $table->unsignedBigInteger('product_id');
             $table->unsignedDecimal('rate', 10, 2);
-            $table->unsignedInteger('qty');
+            $table->float('qty', 10, 2);
             $table->unsignedBigInteger('ref');
             $table->timestamp('date', $precision = 0);
             $table->foreign('product_id')->references('id')->on('products');
